+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs terra-framework-docs
+
 ## 5.35.0 - (September 20, 2021)
 
 * Changed